Move StringStream into Parsely
